# Improve the performance of alignment cops